+++ Examples/UtiliKit-iOSExample/Collection/Cells, Supplementaries and Decorations/Decorations/YellowDecorationView.swift
@@ -1,0 +1,16 @@
+//
+//  YellowDecorationView.swift
+//  UtiliKit
+//
+//  Copyright © 2021 Bottle Rocket Studios. All rights reserved.
+//
+
+import UIKit
+
+class YellowDecorationView: UICollectionReusableView {
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        backgroundColor = .yellow
+    }
+}