+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ##### Enhancements
 
-* None
+* Improve support for registering supplementary and decoration views with UICollectionView.
+[Will McGinty](https://github.com/willmcginty)
+[#110](https://github.com/BottleRocketStudios/iOS-UtiliKit/pull/110)
 
 ##### Bug Fixes
 
@@ -12,7 +14,7 @@
 
 ##### Enhancements
 
-* Make it possible to override functions in container, which aid in `ManagedChild` managemet.
+* Make it possible to override functions in container, which aid in `ManagedChild` management.
 [Will McGinty](https://github.com/willmcginty)
 [#87](https://github.com/BottleRocketStudios/iOS-UtiliKit/pull/87)